% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_graphs.R
\name{eda_graphs}
\alias{eda_graphs}
\title{Insights on the variables through graphs for analysis.}
\usage{
eda_graphs(data, variables = c(1:ncol(data)), filepath = getwd())
}
\arguments{
\item{(dataframe, }{columns, filepath)}
}
\value{

}
\description{
This packages find numeric and categorical variables. Then plots the distribution of values/ levels. Boxplots, Histograms, Pie Charts, Density plots are included.
}
\examples{
eda_graphs(cars,c(1:5),filepath="../folder")

}
